+++ CSK_Module_FTPClient/scripts/Communication/FTPClient/helper/checkAPIs.lua
@@ -1,0 +1,43 @@
+---@diagnostic disable: undefined-global, redundant-parameter, missing-parameter
+
+-- Load all relevant APIs for this module
+--**************************************************************************
+
+local availableAPIs = {}
+
+local function loadAPIs()
+  CSK_FTPClient = require 'API.CSK_FTPClient'
+
+  Container = require 'API.Container'
+  Engine = require 'API.Engine'
+  Engine.AsyncFunction = require 'API.Engine.AsyncFunction'
+
+  Image = require 'API.Image'
+  Log = require 'API.Log'
+  Log.Handler = require 'API.Log.Handler'
+  Log.SharedLogger = require 'API.Log.SharedLogger'
+  Object = require 'API.Object'
+  Timer = require 'API.Timer'
+
+  -- Check if related CSK modules are available to be used
+  local appList = Engine.listApps()
+  for i = 1, #appList do
+    if appList[i] == 'CSK_Module_PersistentData' then
+      CSK_PersistentData = require 'API.CSK_PersistentData'
+    elseif appList[i] == 'CSK_Module_UserManagement' then
+      CSK_UserManagement = require 'API.CSK_UserManagement'
+    end
+  end
+end
+
+local function loadSpecificAPIs()
+  FTPClient = require 'API.FTPClient'
+  Image.Format = {}
+  Image.Format.JPEG = require 'API.Image.Format.JPEG'
+end
+
+availableAPIs.default = xpcall(loadAPIs, debug.traceback) -- TRUE if all default APIs were loaded correctly
+availableAPIs.specific = xpcall(loadSpecificAPIs, debug.traceback) -- TRUE if all specific APIs were loaded correctly
+
+return availableAPIs
+--**************************************************************************